requirements.txt
Pillow
ExifRead
staticmap
pytest
